CalcDataConfirm: refactor -> add utils.ts
